add new js

+++ test.js
@@ -228,7 +228,7 @@
 			*/
 		  }).done(function(msg_r) {
 			alert(msg_r);
-			if(msg_r > 550){
+			if((int)msg_r > 550){
 				return true;
 			}
